--- secretmanager/src/main/java/secretmanager/DestroyRegionalSecretVersion.java
@@ -1,60 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package secretmanager;
-
-// [START secretmanager_destroy_regional_secret_version]
-import com.google.cloud.secretmanager.v1.SecretManagerServiceClient;
-import com.google.cloud.secretmanager.v1.SecretManagerServiceSettings;
-import com.google.cloud.secretmanager.v1.SecretVersion;
-import com.google.cloud.secretmanager.v1.SecretVersionName;
-import java.io.IOException;
-
-public class DestroyRegionalSecretVersion {
-
-  public static void destroySecretVersion() throws IOException {
-    // TODO(developer): Replace these variables before running the sample.
-    String projectId = "your-project-id";
-    String locationId = "your-location-id";
-    String secretId = "your-secret-id";
-    String versionId = "your-version-id";
-    destroyRegionalSecretVersion(projectId, locationId, secretId, versionId);
-  }
-
-  // Destroy an existing secret version.
-  public static void destroyRegionalSecretVersion(
-      String projectId, String locationId, String secretId, String versionId)
-      throws IOException {
-    // Initialize client that will be used to send requests. This client only needs to be created
-    // once, and can be reused for multiple requests. After completing all of your requests, call
-    // the "close" method on the client to safely clean up any remaining background resources.
-    String apiEndpoint = String.format("secretmanager.%s.rep.googleapis.com:443", locationId);
-    SecretManagerServiceSettings secretManagerServiceSettings =
-        SecretManagerServiceSettings.newBuilder().setEndpoint(apiEndpoint).build();
-    try (SecretManagerServiceClient client = 
-        SecretManagerServiceClient.create(secretManagerServiceSettings)) {
-      // Build the name from the version.
-      SecretVersionName secretVersionName = 
-          SecretVersionName.ofProjectLocationSecretSecretVersionName(
-          projectId, locationId, secretId, versionId);
-
-      // Destroy the secret version.
-      SecretVersion version = client.destroySecretVersion(secretVersionName);
-      System.out.printf("Destroyed regional secret version %s\n", version.getName());
-    }
-  }
-}
-// [END secretmanager_destroy_regional_secret_version]
