+++ secretmanager/src/main/java/secretmanager/regionalsamples/DestroyRegionalSecretVersionWithEtag.java
@@ -14,57 +14,59 @@
  * limitations under the License.
  */
 
-package secretmanager;
+package secretmanager.regionalsamples;
 
-// [START secretmanager_update_regional_secret]
-import com.google.cloud.secretmanager.v1.Secret;
+// [START secretmanager_destroy_regional_secret_version_with_etag]
+import com.google.cloud.secretmanager.v1.DestroySecretVersionRequest;
 import com.google.cloud.secretmanager.v1.SecretManagerServiceClient;
 import com.google.cloud.secretmanager.v1.SecretManagerServiceSettings;
-import com.google.cloud.secretmanager.v1.SecretName;
-import com.google.protobuf.FieldMask;
-import com.google.protobuf.util.FieldMaskUtil;
+import com.google.cloud.secretmanager.v1.SecretVersion;
+import com.google.cloud.secretmanager.v1.SecretVersionName;
 import java.io.IOException;
 
-public class UpdateRegionalSecret {
+public class DestroyRegionalSecretVersionWithEtag {
 
-  public static void updateRegionalSecret() throws IOException {
+  public static void destroyRegionalSecretVersion() throws IOException {
     // TODO(developer): Replace these variables before running the sample.
     String projectId = "your-project-id";
     String locationId = "your-location-id";
     String secretId = "your-secret-id";
-    updateRegionalSecret(projectId, locationId, secretId);
+    String versionId = "your-version-id";
+    // Including the quotes is important.
+    String etag = "\"1234\"";
+    destroyRegionalSecretVersion(projectId, locationId, secretId, versionId, etag);
   }
 
-  // Update an existing secret.
-  public static void updateRegionalSecret(
-      String projectId, String locationId, String secretId) 
+  // Destroy an existing secret version.
+  public static void destroyRegionalSecretVersion(
+      String projectId, String locationId, String secretId, String versionId, String etag)
       throws IOException {
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests. After completing all of your requests, call
     // the "close" method on the client to safely clean up any remaining background resources.
+    
+    // Endpoint to call the regional secret manager sever
     String apiEndpoint = String.format("secretmanager.%s.rep.googleapis.com:443", locationId);
     SecretManagerServiceSettings secretManagerServiceSettings =
         SecretManagerServiceSettings.newBuilder().setEndpoint(apiEndpoint).build();
     try (SecretManagerServiceClient client = 
         SecretManagerServiceClient.create(secretManagerServiceSettings)) {
-      // Build the name.
-      SecretName secretName = 
-          SecretName.ofProjectLocationSecretName(projectId, locationId, secretId);
+      // Build the name from the version.
+      SecretVersionName secretVersionName = 
+          SecretVersionName.ofProjectLocationSecretSecretVersionName(
+          projectId, locationId, secretId, versionId);
 
-      // Build the updated secret.
-      Secret secret =
-          Secret.newBuilder()
-              .setName(secretName.toString())
-              .putLabels("secretmanager", "rocks")
+      // Build the request.
+      DestroySecretVersionRequest request =
+          DestroySecretVersionRequest.newBuilder()
+              .setName(secretVersionName.toString())
+              .setEtag(etag)
               .build();
 
-      // Build the field mask.
-      FieldMask fieldMask = FieldMaskUtil.fromString("labels");
-
-      // Update the secret.
-      Secret updatedSecret = client.updateSecret(secret, fieldMask);
-      System.out.printf("Updated regional secret %s\n", updatedSecret.getName());
+      // Destroy the secret version.
+      SecretVersion version = client.destroySecretVersion(request);
+      System.out.printf("Destroyed regional secret version %s\n", version.getName());
     }
   }
 }
-// [END secretmanager_update_regional_secret]
+// [END secretmanager_destroy_regional_secret_version_with_etag]
