+++ secretmanager/src/main/java/secretmanager/regionalsamples/DisableRegionalSecretVersionWithEtag.java
@@ -14,47 +14,59 @@
  * limitations under the License.
  */
 
-package secretmanager;
+package secretmanager.regionalsamples;
 
-// [START secretmanager_enable_regional_secret_version]
+// [START secretmanager_disable_regional_secret_version_with_etag]
+import com.google.cloud.secretmanager.v1.DisableSecretVersionRequest;
 import com.google.cloud.secretmanager.v1.SecretManagerServiceClient;
 import com.google.cloud.secretmanager.v1.SecretManagerServiceSettings;
 import com.google.cloud.secretmanager.v1.SecretVersion;
 import com.google.cloud.secretmanager.v1.SecretVersionName;
 import java.io.IOException;
 
-public class EnableRegionalSecretVersion {
+public class DisableRegionalSecretVersionWithEtag {
 
-  public static void enableRegionalSecretVersion() throws IOException {
+  public static void disableRegionalSecretVersion() throws IOException {
     // TODO(developer): Replace these variables before running the sample.
     String projectId = "your-project-id";
     String locationId = "your-location-id";
     String secretId = "your-secret-id";
     String versionId = "your-version-id";
-    enableRegionalSecretVersion(projectId, locationId, secretId, versionId);
+    // Including the quotes is important.
+    String etag = "\"1234\"";
+    disableRegionalSecretVersion(projectId, locationId, secretId, versionId, etag);
   }
 
-  // Enable an existing secret version.
-  public static void enableRegionalSecretVersion(
-      String projectId, String locationId, String secretId, String versionId)
+  // Disable an existing secret version.
+  public static void disableRegionalSecretVersion(
+      String projectId, String locationId, String secretId, String versionId, String etag)
       throws IOException {
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests. After completing all of your requests, call
     // the "close" method on the client to safely clean up any remaining background resources.
+    
+    // Endpoint to call the regional secret manager sever
     String apiEndpoint = String.format("secretmanager.%s.rep.googleapis.com:443", locationId);
     SecretManagerServiceSettings secretManagerServiceSettings =
         SecretManagerServiceSettings.newBuilder().setEndpoint(apiEndpoint).build();
     try (SecretManagerServiceClient client = 
         SecretManagerServiceClient.create(secretManagerServiceSettings)) {
       // Build the name from the version.
-      SecretVersionName secretVersionName =
-          SecretVersionName.ofProjectLocationSecretSecretVersionName(
-          projectId, locationId, secretId, versionId);
+      SecretVersionName secretVersionName 
+            = SecretVersionName.ofProjectLocationSecretSecretVersionName(
+            projectId, locationId, secretId, versionId);
 
-      // Enable the secret version.
-      SecretVersion version = client.enableSecretVersion(secretVersionName);
-      System.out.printf("Enabled regional secret version %s\n", version.getName());
+      // Build the request.
+      DisableSecretVersionRequest request =
+          DisableSecretVersionRequest.newBuilder()
+              .setName(secretVersionName.toString())
+              .setEtag(etag)
+              .build();
+
+      // Disable the secret version.
+      SecretVersion version = client.disableSecretVersion(request);
+      System.out.printf("Disabled regional secret version %s\n", version.getName());
     }
   }
 }
-// [END secretmanager_enable_regional_secret_version]
+// [END secretmanager_disable_regional_secret_version_with_etag]
