+++ secretmanager/src/main/java/secretmanager/regionalsamples/EnableRegionalSecretVersionWithEtag.java
@@ -14,63 +14,59 @@
  * limitations under the License.
  */
 
-package secretmanager;
+package secretmanager.regionalsamples;
 
-// [START secretmanager_list_regional_secret_versions_with_filter]
-import com.google.cloud.secretmanager.v1.ListSecretVersionsRequest;
+// [START secretmanager_enable_regional_secret_version_with_etag]
+import com.google.cloud.secretmanager.v1.EnableSecretVersionRequest;
 import com.google.cloud.secretmanager.v1.SecretManagerServiceClient;
-import com.google.cloud.secretmanager.v1.SecretManagerServiceClient.ListSecretVersionsPagedResponse;
 import com.google.cloud.secretmanager.v1.SecretManagerServiceSettings;
-import com.google.cloud.secretmanager.v1.SecretName;
+import com.google.cloud.secretmanager.v1.SecretVersion;
+import com.google.cloud.secretmanager.v1.SecretVersionName;
 import java.io.IOException;
 
-public class ListRegionalSecretVersionsWithFilter {
+public class EnableRegionalSecretVersionWithEtag {
 
-  public static void listRegionalSecretVersions() throws IOException {
+  public static void enableRegionalSecretVersion() throws IOException {
     // TODO(developer): Replace these variables before running the sample.
     String projectId = "your-project-id";
     String locationId = "your-location-id";
     String secretId = "your-secret-id";
-    // Follow https://cloud.google.com/secret-manager/docs/filtering
-    // for filter syntax and examples.
-    String filter = "create_time>2021-01-01T00:00:00Z";
-    listRegionalSecretVersions(projectId, locationId, secretId, filter);
+    String versionId = "your-version-id";
+    // Including the quotes is important.
+    String etag = "\"1234\"";
+    enableRegionalSecretVersion(projectId, locationId, secretId, versionId, etag);
   }
 
-  // List all secret versions for a secret.
-  public static void listRegionalSecretVersions(
-      String projectId, String locationId, String secretId, String filter)
+  // Enable an existing secret version.
+  public static void enableRegionalSecretVersion(
+      String projectId, String locationId, String secretId, String versionId, String etag)
       throws IOException {
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests. After completing all of your requests, call
     // the "close" method on the client to safely clean up any remaining background resources.
+    
+    // Endpoint to call the regional secret manager sever
     String apiEndpoint = String.format("secretmanager.%s.rep.googleapis.com:443", locationId);
     SecretManagerServiceSettings secretManagerServiceSettings =
         SecretManagerServiceSettings.newBuilder().setEndpoint(apiEndpoint).build();
     try (SecretManagerServiceClient client = 
         SecretManagerServiceClient.create(secretManagerServiceSettings)) {
-      // Build the parent name.
-      SecretName secretName = 
-          SecretName.ofProjectLocationSecretName(projectId, locationId, secretId);
+      // Build the name from the version.
+      SecretVersionName secretVersionName = 
+          SecretVersionName.ofProjectLocationSecretSecretVersionName(
+          projectId, locationId, secretId, versionId);
 
-      // Get filtered versions.
-      ListSecretVersionsRequest request =
-          ListSecretVersionsRequest.newBuilder()
-              .setParent(secretName.toString())
-              .setFilter(filter)
+      // Build the request.
+      EnableSecretVersionRequest request =
+          EnableSecretVersionRequest.newBuilder()
+              .setName(secretVersionName.toString())
+              .setEtag(etag)
               .build();
 
-      ListSecretVersionsPagedResponse pagedResponse = client.listSecretVersions(request);
-
-      // List all versions and their state.
-      pagedResponse
-          .iterateAll()
-          .forEach(
-              version -> {
-                System.out.printf("Regional secret version %s, %s\n", 
-                    version.getName(), version.getState());
-              });
+      // Enable the secret version.
+      SecretVersion version = client.enableSecretVersion(request);
+      System.out.printf("Enabled regional secret version %s\n", version.getName());
     }
   }
 }
-// [END secretmanager_list_regional_secret_versions_with_filter]
+// [END secretmanager_enable_regional_secret_version_with_etag]
