+++ secretmanager/src/main/java/secretmanager/regionalsamples/UpdateRegionalSecret.java
@@ -14,67 +14,59 @@
  * limitations under the License.
  */
 
-package secretmanager;
+package secretmanager.regionalsamples;
 
-// [START secretmanager_iam_revoke_access_with_regional_secret]
+// [START secretmanager_update_regional_secret]
+import com.google.cloud.secretmanager.v1.Secret;
 import com.google.cloud.secretmanager.v1.SecretManagerServiceClient;
 import com.google.cloud.secretmanager.v1.SecretManagerServiceSettings;
 import com.google.cloud.secretmanager.v1.SecretName;
-import com.google.iam.v1.Binding;
-import com.google.iam.v1.GetIamPolicyRequest;
-import com.google.iam.v1.Policy;
-import com.google.iam.v1.SetIamPolicyRequest;
+import com.google.protobuf.FieldMask;
+import com.google.protobuf.util.FieldMaskUtil;
 import java.io.IOException;
 
-public class IamRevokeAccessWithRegionalSecret {
+public class UpdateRegionalSecret {
 
-  public static void iamRevokeAccessWithRegionalSecret() throws IOException {
+  public static void updateRegionalSecret() throws IOException {
     // TODO(developer): Replace these variables before running the sample.
     String projectId = "your-project-id";
     String locationId = "your-location-id";
     String secretId = "your-secret-id";
-    String member = "user:foo@example.com";
-    iamRevokeAccessWithRegionalSecret(projectId, locationId, secretId, member);
+    updateRegionalSecret(projectId, locationId, secretId);
   }
 
-  // Revoke a member access to a particular secret.
-  public static void iamRevokeAccessWithRegionalSecret(
-      String projectId, String locationId, String secretId, String member)
+  // Update an existing secret.
+  public static void updateRegionalSecret(
+      String projectId, String locationId, String secretId) 
       throws IOException {
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests. After completing all of your requests, call
     // the "close" method on the client to safely clean up any remaining background resources.
+    
+    // Endpoint to call the regional secret manager sever
     String apiEndpoint = String.format("secretmanager.%s.rep.googleapis.com:443", locationId);
     SecretManagerServiceSettings secretManagerServiceSettings =
         SecretManagerServiceSettings.newBuilder().setEndpoint(apiEndpoint).build();
     try (SecretManagerServiceClient client = 
         SecretManagerServiceClient.create(secretManagerServiceSettings)) {
-      // Build the name from the version.
+      // Build the name.
       SecretName secretName = 
           SecretName.ofProjectLocationSecretName(projectId, locationId, secretId);
 
-      // Request the current IAM policy.
-      Policy policy =
-          client.getIamPolicy(
-              GetIamPolicyRequest.newBuilder().setResource(secretName.toString()).build());
+      // Build the updated secret.
+      Secret secret =
+          Secret.newBuilder()
+              .setName(secretName.toString())
+              .putLabels("secretmanager", "rocks")
+              .build();
 
-      // Search through bindings and remove matches.
-      String roleToFind = "roles/secretmanager.secretAccessor";
-      for (Binding binding : policy.getBindingsList()) {
-        if (binding.getRole() == roleToFind && binding.getMembersList().contains(member)) {
-          binding.getMembersList().remove(member);
-        }
-      }
+      // Build the field mask.
+      FieldMask fieldMask = FieldMaskUtil.fromString("labels");
 
-      // Save the updated IAM policy.
-      client.setIamPolicy(
-          SetIamPolicyRequest.newBuilder()
-              .setResource(secretName.toString())
-              .setPolicy(policy)
-              .build());
-
-      System.out.printf("Updated IAM policy for %s\n", secretId);
+      // Update the secret.
+      Secret updatedSecret = client.updateSecret(secret, fieldMask);
+      System.out.printf("Updated regional secret %s\n", updatedSecret.getName());
     }
   }
 }
-// [END secretmanager_iam_revoke_access_with_regional_secret]
+// [END secretmanager_update_regional_secret]
